Move error tests into tests/errors/
Consolidate exception-raising tests under tests/errors/, matching the
convention that error contracts (no compilable golden-file surface) get
pytest unit coverage there.

- Relocate tests/test_record_naming.py to
  tests/errors/test_record_naming_errors.py (all InvalidRecordNameError).
- Split tests/test_variables.py: IncompatibleFormatsError /
  NotImplementedError cases to tests/errors/
  test_incompatible_format_errors.py; the lone Rust rendering test
  moves to tests/test_languages.py. Delete test_variables.py.
- Extract the two error tests from tests/test_languages.py into
  tests/errors/test_null_in_collection_errors.py and
  tests/errors/test_special_float_errors.py; drop now-unused imports.

Test-only reorganization; no behavior change. 205 tests pass.
